Refactor smoke example to hackable shader example
Replaces the smoke-example with a new hackable-example, moving and renaming core files, updating the shader implementation, and introducing a more flexible, quality-configurable shader system. Adds new config, vertex, and fragment shader files, updates the main App to use HackableExample, and removes the old smoke-example components. The new structure supports easier customization and improved rendering control.
