+++ templates/shader/src/App.tsx
@@ -1,10 +1,11 @@
 import { FluidExample } from './fluid-example/FluidExample'
+import { HackableExample } from './hackable-example/HackableExample'
 import './shader.css'
 
 function App() {
 	return (
 		<div className="shader-app">
-			{/* <SmokeExample /> */}
+			<HackableExample />
 			<FluidExample />
 		</div>
 	)
