+++ templates/shader/src/hackable-example/README.md
@@ -1,0 +1,3 @@
+# Hackable example
+
+This example is a simple, hackable shader example.
